Utility to calculate dates in past or future
There is surely a better way to do this in Go, but this is what I
came up with that works. Takes a user-supplied date, then asks for
what values to take from or add to that date. Supplies the new
calulated date in the same format as the original entry. Useful
for my genealogy work.